requirements.txt
diffusers==0.33.1
embedding_reader==1.8.1
fire==0.7.1
fsspec==2024.3.1
magma==0.3.9
matplotlib==3.8.4
multidict==6.0.5
nudenet==3.4.2
numpy>=1.26.0
pandas
Pillow==12.1.0
scikit_learn==1.4.2
scipy==1.17.0
seaborn==0.13.2
torch==2.5.1
torchmetrics==1.8.2
torchvision==0.20.1
tqdm==4.66.4
transformers==4.41.2
wordcloud==1.9.6
